Switch from picking up VotingAreas each time from latest data to fixed GeoJson
Our TaskManager has list of tasks with geoemetries and Ids as they were in Feburary of 2022, but since VLV.shp keeps changing with time we started diverging on Geoemetries and IDs of loaded areas in JOSM vs. taks, hence use VLV.geojson from Februrary...

+++ OsmGursBuildingImport/Program.cs
@@ -33,7 +33,8 @@ namespace OsmGursBuildingImport
 
             var dataFolder = Path.Combine(repoRoot, "data");
             var tempDir = Path.Combine(dataFolder, "temp/");
-            Directory.Delete(tempDir, true);
+            if (Directory.Exists(tempDir))
+                Directory.Delete(tempDir, true);
 
             await Process.Start(new ProcessStartInfo()
             {
